pyproject.toml
[tool.poetry]
name = "test-sgs"
version = "0.1.0"
description = ""
authors = ["polenom <vitalimitsevich@yandex.by>"]
readme = "README.md"
packages = [{include = "test_sgs"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "5.0.1"
django-filter = "23.5"
pillow = "10.2.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
